+++ NFlags.Tests/DataTypes/EmptyArgumentsType.cs
@@ -1,0 +1,6 @@
+namespace NFlags.Tests.DataTypes
+{
+    public class EmptyArgumentsType
+    {
+    }
+}