+++ NFlags.Tests/DataTypes/NotArrayParameterSeries.cs
@@ -1,0 +1,8 @@
+namespace NFlags.Tests.DataTypes
+{
+    public class NotArrayParameterSeries
+    {
+        [ParameterSeries("", "")]
+        public int ParameterSeries;        
+    }
+}