+++ NFlags.Tests/DataTypes/OptionWithoutSetter.cs
@@ -1,0 +1,8 @@
+namespace NFlags.Tests.DataTypes
+{
+    public class OptionWithoutSetter
+    {
+        [Option("", "", "")]
+        public int Option { get; }    
+    }
+}