+++ NFlags/GenericCommandExtension/CommandRegisterer.cs
@@ -15,6 +15,9 @@ namespace NFlags.GenericCommandExtension
 
         public void RegisterParameterSeries(MemberInfo member, ParameterSeriesAttribute parameterSeriesAttribute)
         {
+            ValidatePropertySetter(member);
+            ValidateParameterSeriesMemberType(member);
+
             var method = typeof(CommandConfigurator).GetMethods().Single(m => m.IsGenericMethod && m.Name == "RegisterParameterSeries");
             var generic = method.MakeGenericMethod(TypeHelper.GetMemberType(member).GetElementType());
             generic.Invoke(_commandConfigurator, new object[] {parameterSeriesAttribute.Name, parameterSeriesAttribute.Description});
@@ -22,6 +25,8 @@ namespace NFlags.GenericCommandExtension
 
         public void RegisterParameter(MemberInfo member, ParameterAttribute parameterAttribute)
         {
+            ValidatePropertySetter(member);
+                
             var method = typeof(CommandConfigurator).GetMethod("RegisterParam");
             var generic = method.MakeGenericMethod(TypeHelper.GetMemberType(member));
             generic.Invoke(_commandConfigurator, new[] {parameterAttribute.Name, parameterAttribute.Description, parameterAttribute.DefaultValue});
@@ -29,6 +34,9 @@ namespace NFlags.GenericCommandExtension
 
         public void RegisterFlag(MemberInfo member, FlagAttribute flagAttribute)
         {
+            ValidatePropertySetter(member);
+            ValidateFlagMemberType(member);
+
             if (flagAttribute.Abr != null)
                 _commandConfigurator.RegisterFlag(flagAttribute.Name, flagAttribute.Abr, flagAttribute.Description, flagAttribute.DefaultValue);
             else
@@ -37,6 +45,8 @@ namespace NFlags.GenericCommandExtension
 
         public void RegisterOption(MemberInfo member, OptionAttribute optionAttribute)
         {
+            ValidatePropertySetter(member);
+            
             var parametersCount = optionAttribute.Abr != null ? 4 : 3;
             var method = typeof(CommandConfigurator).GetMethods()
                 .Single(info => info.Name == "RegisterOption" && info.IsGenericMethod && info.GetParameters().Length == parametersCount);
@@ -61,6 +71,27 @@ namespace NFlags.GenericCommandExtension
 
             var generic = method.MakeGenericMethod(TypeHelper.GetMemberType(member));
             generic.Invoke(_commandConfigurator, parameters);
+        }
+
+        private static void ValidatePropertySetter(MemberInfo member)
+        {
+            var propertyInfo = member as PropertyInfo;
+            if (propertyInfo != null && propertyInfo.GetSetMethod() == null)
+                throw new PropertyWithoutSetterException(member.Name);
+        }
+
+        private static void ValidateParameterSeriesMemberType(MemberInfo member)
+        {
+            var memberType = TypeHelper.GetMemberType(member);
+            if (!memberType.IsArray)
+                throw new IncorrectParameterSeriesMemberTypeException(member.Name, memberType);
+        }
+
+        private static void ValidateFlagMemberType(MemberInfo member)
+        {
+            var memberType = TypeHelper.GetMemberType(member);
+            if (memberType != typeof(bool))
+                throw new IncorrectFlagMemberTypeException(member.Name, memberType);
         }
     }
 }