+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 1.4.0
  * Generic custom args command registration
- * Method RegisterSubcommand marked as obsolete and RegisterCommand method introduced.
+ * Method RegisterSubcommand marked as obsolete and RegisterCommand method introduced
+ * Introduced PrintHelpOnExecute
+ * Introduced default commands 
 
 # 1.3.1
  * generic options and parameters getters in CommandArgs